✨feat : 한번에 저장하기 Promise.all 사용하여 동기적으로 처리
