+++ src/components/PostResult/PostButton.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import ButtonWithTip from '../../components/common/Button/ButtonWithTip/ButtonWithTip';
 import { useState } from 'react';
-import { isAndroid } from '../../utils/utils';
+import { copyText, downloadImage, isAndroid } from '../../utils/utils';
 
 interface PostButtonProps {
   image: string;
@@ -12,49 +12,32 @@ export default function PostButton({ image, text }: PostButtonProps) {
   // isSaved - true, [공유하기] show
   const [isSaved, setIsSaved] = useState(false);
 
-  function saveImage() {
-    if (isAndroid()) {
-      return Android.downloadImage(image);
-    } else {
-      fetch(image, {
-        method: 'GET',
-      })
-        .then((res) => {
-          return res.blob();
-        })
-        .then((blob) => {
-          const blobURL = URL.createObjectURL(blob);
-
-          const aTag = document.createElement('a');
-          aTag.href = blobURL;
-          aTag.download = 'sodong_image.png';
-          aTag.click();
-
-          alert('다운로드 성공!');
-        })
-        .catch((e) => {
-          console.error(e);
-          alert(e);
-        });
-    }
-  }
-
-  function copyText() {
-    try {
-      if (!navigator?.clipboard?.writeText)
-        throw new Error('복사 기능이 제공되지 않는 브라우저입니다.');
-      window.navigator.clipboard.writeText(text).then(() => {
-        alert('복사 성공!');
-      });
-    } catch (e) {
-      alert(e);
-    }
-  }
+  // (1) [한번에 저장하기]
+  // 텍스트 복사 -> 이미지 저장 -> 모두 성공 시, isSaved!
   function handleSaveAll() {
-    setIsSaved(true);
-    saveImage();
-    copyText();
+    const saveFunc = isAndroid() ? saveImage : downloadImage;
+    console.log('>>> Func', saveFunc);
+    Promise.all([saveFunc(image), copyText(text)])
+      .then((res) => {
+        console.log('>> res'), res;
+        alert('성공하였습니다');
+        setIsSaved(true);
+      })
+      .catch((error) => {
+        alert(error);
+      });
   }
+
+  function saveImage() {
+    return new Promise((resolve, reject) => {
+      const uri = Android.downloadImage(image);
+      if (uri === '') reject();
+      else resolve(uri);
+    });
+  }
+
+  // (2) SNS 공유하기
+  // Android인 경우, openApp() / else인 경우, 스토어로 랜딩
   function handleShare() {}
 
   return (
