+++ src/pages/PostResult/PostResult.tsx
@@ -2,66 +2,14 @@ import styled from 'styled-components';
 import PostTitle from '../../components/PostResult/PostTitle';
 import PostImage from '../../components/PostResult/PostImage';
 import PostText from '../../components/PostResult/PostText';
-import { useEffect, useState } from 'react';
 import PostButton from '../../components/PostResult/PostButton';
-import { isAndroid } from '../../utils/utils';
 
-interface Post {
-  image: string;
-  text: string;
-}
 export default function PostResult() {
-  const [post, setPost] = useState<Post>({ image: '', text: '' });
-  useEffect(() => {
-    fetch('/api/postings/123')
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        setPost(data);
-      });
-  }, []);
-
-  //// TEST START ////
-  function copyText() {
-    try {
-      if (!navigator?.clipboard?.writeText)
-        throw new Error('복사 기능이 제공되지 않는 브라우저입니다.');
-      // 클립보드에 복사
-      window.navigator.clipboard.writeText('dhkfkfkk').then(() => {
-        alert('copy');
-      });
-    } catch (e) {
-      alert(e);
-    }
-  }
-
-  function saveImage() {
-    if (isAndroid()) {
-      return Android.downloadImage(post.image);
-    } else {
-      fetch(post.image, {
-        method: 'GET',
-      })
-        .then((res) => {
-          return res.blob();
-        })
-        .then((blob) => {
-          const blobURL = URL.createObjectURL(blob);
-
-          const aTag = document.createElement('a');
-          aTag.href = blobURL;
-          aTag.download = 'sodong_image.png';
-          aTag.click();
-
-          alert('다운로드 성공!');
-        })
-        .catch((e) => {
-          console.error(e);
-          alert(e);
-        });
-    }
-  }
-  //// TEST END ////
+  const post = {
+    text: '망원동 성공 맛집 딸기프토 감성 카페☕️ \n쫀득, 부드러운 토스트 위에 크림치즈와 딸기가 듬뿍!🍓 \n비주얼부터 맛까지 빠짐없는 망원동 필수 코스 ✨\n\n# 망원동카페 # 합정카페 \n🏷️ 망원동 #콘웰 \n📍 서울 마포구 월드컵로15길 40 2층',
+    image:
+      'https://firebasestorage.googleapis.com/v0/b/twitter-72580.appspot.com/o/img%2Fimage_example.png?alt=media&token=5cf53d9e-93e2-4234-a075-8261c87df3ea',
+  };
 
   return (
     <PostResultContainter>
@@ -69,11 +17,6 @@ export default function PostResult() {
       <PostImage url={post.image} />
       <PostText text={post.text} />
       <PostButton image={post.image} text={post.text} />
-      <div onClick={copyText}>
-        Copy
-        <br /> <br />
-      </div>
-      <div onClick={saveImage}>Download</div>
     </PostResultContainter>
   );
 }
